\name{LevelSatPhd}
\alias{LevelSatPhd}
\docType{data}
\title{
Questionnaire Data
}
\description{
The sample data, as part of a large survey, corresponds to 10 people who have the PhD degree and shows the level of satisfaction about some issues.
}
\usage{data(LevelSatPhd)}
\format{
  This data frame contains 10 observation for the following 5 ordinal variables,
  with four categories each:
  (1= "Very Satisfied", 2= "Somewhat Satisfied",3="Somewhat dissatisfied",
  4="Very dissatisfied")
  \describe{
    \item{Salary}{}
    \item{Benefits}{}
    \item{Job Security}{}
    \item{Job Location}{}
    \item{Working conditions}{}
  }
}
\source{
Spanish Statistical Institute. Survey of PDH holders, 2006. 
URL: http://www.ine.es.
}
\examples{
data(LevelSatPhd)
}
\keyword{datasets}
